requirements.txt
fastapi
uvicorn[standard]
databases[aiosqlite]
python-dotenv
sqlalchemy
datetime
greenlet
pydantic-settings
mypy
